app/requirements.txt
flask
scrapy
flask-cors
pyyaml